Add Options and SetFiles classes and constants module
