+++ constants.rb
@@ -1,0 +1,5 @@
+module Constants
+  FILE_NAMES = ['books.json', 'people.json', 'rentals.json']
+  DIRECTORY_PATH = File.join(Dir.pwd, 'JSON_files')
+end
+
